hack around alignment issues which crop up on platforms less tolerant of such things than x86 is.  Switch from long to int for types which should be 32bit.  Can't use uint32_t as win32 does not have it.  Now compiles and works on solaris sparc 32 and 64 bit
git-svn-id: svn://svn.code.sf.net/p/tivodecode/code/tivodecode/trunk@13 3cae7af1-9334-450b-b769-427e0023cdf8

+++ sha1.h
@@ -2,8 +2,8 @@
 #define SHA1_H_
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    unsigned int state[5];
+    unsigned int count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
